feat: Add a post for log collection (part 1)
This establishes an environment to develop within, including a Cargo
project and validating access to kubectl and a docker registry.

The Cargo project, monitoring-rs, produces a single binary, also
monitoring-rs, that currently just periodically prints the file tree
under /var/log. The post goes on to validate that, indeed, this will
allow us to access Kubernetes container logs.