feat: Initial attempt at log collector
This was a bit of a disaster. The decision to use `notify` caused a bit
of a blocker whereby the local behaviour was unexpected. This turned out
to be a known issue (https://github.com/notify-rs/notify/issues/240)
with notify, related to how fsevents works (it sends all events when the
file closes).

A workaround of running the container through Docker was persued, but
the current set up for that is not very smooth, so if that's the way it
has to be we should dedicate some words to setting up a decent harness
for experimentation.

However, this was semi-successful in that in brought us to the meat of
the log watcher logic. This should make it possible to take a step back
and try again with a clearer idea of what we want, and hopefully make it
to a better outcome!