feat: Brute-force end-to-end implementation
This post builds on the restructuring of the last one by "brute-forcing"
an inefficient and untidy end-to-end implementation. The result is a
binary that runs the collector loop in a thread and stores collected log
entries in a separate data file. In another thread runs a web server
with a `/logs/*key` endpoint that can be used to read log entries stored
in those data files.