feat: Add a post 'consolidating' the end-to-end implementation
This is mostly a stop-gap post to tidy up some of the 'snagging' issues
from driving towards the end-to-end solution. Primarily, this involved
optimising the `docker build`, adding some tests to `api`, and tidying
up `main()`.

In the process of adding tests to `log_collector` we discover that the
project cannot build natively due to the dependency on inotify. We make
this more obvious by turning inotify into a linux-only dependency so
that `cargo check` now fails locally, as well as using some conditional
compilation shenanigans to give us an obvious error and allow tests to
still run.