feat: Support `key=value` metadata for logs
This is another meaty post, in which support has been added for log
collector to annotate log entries with key-value metadata. The database
has been updated to persist this metadata, and allow retrieving logs
by key-value. The API has been updated to require key and value to be
supplied in the path.

The original intention was to explore pulling metadata from Kubernetes,
but starting here would have required working out how to stub the
Kubernetes API locally/in tests. So instead the focus was shifted to
refactoring to support metadata through the collector, storage, and api
modules.

A future post will be able to focus on building a
log_collector::kubernetes module that interacts with the API.