feat: Add a post with a 2nd attempt at a kubernetes collector
This brings the implementation closer to featuring a a Kubernetes log
collector, but as yet "no cigar". We have introduced some additional
'infrastructure' to allow selecting from multiple log collectors in
future, before getting bogged down in several unexpected issues related
to differences between `inotify` and `kqueue`, and handling symlinks.

In the end, we've boosted the documentation for `watcher` to include
"callee responsibilities" to ensure consistent behaviour across
platforms, and introduced new tests (and a mock watcher implementation)
to exercise these responsibilities.

Hopefully it will be 3rd time lucky for the Kubernetes collector!