feat: Create an initial kubernetes log collector
This builds on the past few rounds of refactoring to finally introduce a
working, albeit slightly hacky, Kubernetes log collector. This wraps a
directory collector, but features a default root path, and augments
collected log entries with the labels from the host Kubernetes pod.

This also features some refactoring of the `log_collector::watcher`
module to better-separate the implementations.