feat: Add a post with some rambling and some deployment
This should probably have been two posts. We started off with a rambling
discussion about possible architectural directions for separating log
files from log entries/collectors.

We then abruptly stopped doing that and:

- Added some caching to the Kubernetes collector.
- Added a `make` task to deploy to Kubernetes.
- Debugged and fixed the Kubernetes deployment.

The last part turned into a bit of an odyssey thanks to ctz/rustl#184,
which forced us to switch to openssl for TLS which lead us down a rabbit
hole of static/dynamic linking shenanigans. We made it eventually,
however, and now have a functional log collector 'soaking' in a real
Kubernetes cluster 🎉