feat: Add a post discussing next steps for log collection
This is mostly expanding on various GitHub issues created to capture
some of the outstanding work to 'productionise' log collection. After
covering them all, we arrive at the conclusion that adding persistence
to what we have is enough to put a pin in log collection for now, and
start looking at metric collection!

Fixes #12.