feat: Add a post kicking off a new log/metrics database
This post was a bit meandering. Who'd have thought database design was a
pretty deep area with a lot of complex techniques. The post covers some
analysis of Elasticsearch, and substantially more of Prometheus, that
didn't really go anywhere (but I learned a lot).

In the end we kicked off a new implementation in `database` that
actually covers the same functionality as `log_database` but in a much
more naive way – it's basically just a `Vec<(Labels, Event)>` that is
written to a JSON file in `drop`.

The proposed next step is to define some representative scenarios and
use them to drive future development.