feat: Add a post about load testing and more database design
This started as an ernest attempt to try and build a database in a
performance-test-driven way. We did manage to build a small `loadgen`
crate that can be configured to simulate write loads with different
characteristics. This could be used to run simulations which could then
be measured and profiled to determine whether performance is sufficient
and if not what the next thing to fix is.

At some point, I started sinking back into DB exploration and so the
post abruptly switches to some very loosely organised notes about
different techniques and technologies, followed by some hopefully useful
insights:

- Key-value stores are a bad fit for time-series *data*, but could work
  for indices (labels).
- Rust has a well-developed Apache Arrow ecosystem, and the InfluxDB IOx
  project is betting on it to build their time-series database to
  replace the existing InfluxDB. Apache Arrow could be used to build an
  embedded time-series database that would suit our needs.